[3] Add WM&SF Properties sub-panel in IME panels
If there are WM or SF trace entries that correspond with IME trace
entries (i.e. the WM/SF entry comes before the IME entry in time),
the 'WM & SF Properties' sub-panel will be displayed below the
'Hierarchy' sub-panel in the IME trace panel.

- This sub-panel currently displays properties extracted from
WM entry, not yet SF entry.
- Code is added in 'ime_processing.js' to extract WM properties,
and 'freeze: false' is set in 'transform_ime.js' so that IME entry
objects are not frozen and can be modified during processing
- SF entry, displayed as a tree in the Hierarchy sub-panel, is pruned
to only ImeContainer or IME-snapshot node and its sibling nodes.

Note: there is a bug in this CL (b/236679852) that is fixed later
in CL[6]. Also, null values are not handled well in this CL, and
will be fixed later in CL[4].

Screenshot: https://screenshot.googleplex.com/6RXiRyX47qLAL9i.png

Bug: 236226833
Test: manual on local winscope

Change-Id: I07be7f492a6275c3fdafd6d1f523fdf4d552d16b