Downgrade vue-material version
Fix vue-material errors that show up when using MdSelect (which
we use for <Overlay />). These errors are a known bug in the
vue-material codebase
(https://github.com/vuematerial/vue-material/issues/2285) but no
patch is out for it yet. Hence, downgrade vue-material to the
latest version that does not have these errors.

Test: manual with local build of Winscope
Bug: 238262637
Change-Id: Icffead3cbf1fa1f2a6e4ec83564078f396e660cf