compare_cts_reports: add the `--ignore-abi` option
When `--ignore-abi` is specified, the comparison tool would
ignore test_abi while aggregating the reports. It means
that two tests would be considered as the same one as long
as they have the same module_name#class_name.test_name.

This option is useful when comparing a x86_64 report with
an arm64 report.

Bug: 315073470
Test: ./compare_cts_reports.py \
        -r x86_64_report.zip -r arm64_report.zip \
	--mode 2 --output-dir output --ignore-abi
Change-Id: Iab04be72676db8c67518e1f7bfbba482e1f971c6