Fix wrong initial timestamp in UI.
TimelineData initialized without knowing first active trace. This caused the initial timestamp to be set as the first timestamp of all loaded traces.
Now the first active trace is set to be the first trace in the viewer display order, discounting screen recording, so the initial timestamp is set as the first timestamp of that active trace.

Bug: b/323130677
Test: npm run test:unit:ci
Change-Id: I6177084b6c4a942603b6d111d44d71d88b9cc387