Remove an invalid file from the repo.
This repo needs to be cloned on Windows, but this file is preventing
that. Remove the file from the repo and temporarily disable the test
that uses it. This code will either need to be moved to a repo that
doesn't need to be used on Windows, or be rewritten to dynamically
create the file with the invalid name.

Bug: None
Test: None
Change-Id: I67aa0934f8b068517ac8b19f7ca75805c35ce828