Stop using persist.sys.time.offset
Seems like it's not always accurate
e.g. in bug below, sys.time.offset is 8 hours but sys.timezone gives offset of 7 hours for timestamps in trace.

This means we have no need to include utcOffsetMs in the TimezoneInfo interface, as we're never going to be receiving it alongside the timezone string.

Fixes: 340198224
Test: npm run test:unit:ci
Change-Id: I020b2f4b189ae6e51bf1b636faec1161ffb13b91