Only stream device aware permissions in VDM demo app
We made a change recently to only allow device aware permissions to be streamed to a virtual device. If requested permissions include non-device aware permissions we show a warning dialog. We need to make corresponding changes in the VDM permission demo app to limit requested permissions to be device aware so that the permission grant dialog can be displayed on a virtual device.

Bug: 324470724
Test: manual
Change-Id: Iebef5dd7513a93f9e6160cac92a50df66feb6f0d