Fix color/alpha and transform formatting.
Ensure color and alpha always reported separately (not alpha inside brackets).
Do not discard alpha when filtering defaults, or color won't be able to format properly (because we reuse the color formatter for rgb colors in Transactions, so we can't just fill in alpha = 0 if it's not there).
Do not discard transform type when filtering defaults for same reason.

Fixes: 350946345
Test: npm run test:unit:ci

Change-Id: Ic510035fc0cd827970655c342cc35cc4892c9e4a