License classifier crate
When run on 377 crates in external/rust/crates, there are 16 crates for
which it is unable to find a license file to satisfy the license terms,
often because the crate does not include a license file.

For the remaining 361 crates, it successfully finds all the required
license files, which is a 96% success rate.

Bug: 339424309
Test: Ran it on external/rust/crates
Change-Id: I1cc92345f04998db0b10a28cee928668d8a854c6