Add current index to CUJs and Transitions ui data.
Getting confusing with large transitions traces regarding why the up/down arrow keys and the left/right arrow keys have different behaviours. Because the table is ordered by id, the left/right arrow keys can cause jumping around on the table, and it's unclear to the user that the up/down arrow keys do not actually update the timeline.

Bug: 360079237
Test: npm run test:unit:ci
Change-Id: I814b492b8d5da47cd9f170cbed907a2f8d797d46