Preserve Cargo.lock when running cargo_embargo.
cargo_embargo will respect a Cargo.lock file if one is present in the
vendored code.

If not, it will create one, which we have been preserving and checking
in, leading to unintended and unhelpful churn.

This change ensures that any existing Cargo.lock files in vendored code
are preserved, but no new Cargo.lock files are introduced.

Change-Id: Iee69e214484faba596995e8e1b813b7ae74fd7d8
Bug: 339424309
Test: treehugger